Custom delete dialog added with delete asset

+++ src/app/modules/assetmate/assetmate-view/Asset/view-asset/view-asset.component.ts
@@ -11,6 +11,8 @@ import { AssetCodeComponent } from '../asset-code/asset-code.component';
 import { ConfirmDialogModel, ConfirmDialogComponent } from '../../../../../shared/confirm-dialog/confirm-dialog.component';
 import { SpinnerService } from '../../../../../public service/spinner.service';
 import { MatSnackBar } from '@angular/material';
+import { DialogService } from '../../../../../public service/dialog.service';
+import { AppDialogData } from '../../../../../model/appDialogData';
 
 @Component({
   selector: 'app-view-asset',
@@ -53,6 +55,7 @@ export class ViewAssetComponent implements AfterViewInit, OnDestroy {
     public dialog: MatDialog,
     private spinnerService: SpinnerService,
     private snackBar: MatSnackBar,
+    private dialogService: DialogService
   ) {
 
   }
@@ -152,24 +155,56 @@ export class ViewAssetComponent implements AfterViewInit, OnDestroy {
 
 
   /*********************************************************** Delete Particular Asset *******************************************************************/
-  deleteAsset(assetId: number) {
-    // alert('are you sure?');
-    const message = `Are you sure you want to do this?`;
-    const dialogData = new ConfirmDialogModel("Confirm Action", message);
-    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
-      maxWidth: "1000px",
-      data: dialogData
-    });
-    dialogRef.afterClosed().subscribe(dialogResult => {
-      this.result = dialogResult;
-    });
-    this.assetmateService.deleteAsset(assetId).subscribe(res => {
-      this.toastr.success(res.message);
-      this.getAllAssets(this.categoryID, this.page);
-    })
-    error => {
-      this.toastr.error(error.message);
+  // deleteAsset(assetId: number) {
+  //   // alert('are you sure?');
+  //   const message = `Are you sure you want to do this?`;
+  //   const dialogData = new ConfirmDialogModel("Confirm Action", message);
+  //   const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+  //     maxWidth: "1000px",
+  //     data: dialogData
+  //   });
+  //   dialogRef.afterClosed().subscribe(dialogResult => {
+  //     this.result = dialogResult;
+  //   });
+  //   this.assetmateService.deleteAsset(assetId).subscribe(res => {
+  //     this.toastr.success(res.message);
+  //     this.getAllAssets(this.categoryID, this.page);
+  //   })
+  //   error => {
+  //     this.toastr.error(error.message);
+  //   }
+  // }
+
+  deleteAsset(assetId: number, assetTitle: string) {
+
+    let appDialogData: AppDialogData = {
+      visibilityStatus: true,
+      title: 'DELETE ASSETE',
+      message: `Are your sure you want to delete assete "${assetTitle}"`,
+      positiveBtnLable: "Yes",
+      negativeBtnLable: "Cancel"
     }
+
+    this.dialogService.setDialogVisibility(appDialogData);
+
+    this.dialogService.getUserDialogAction().subscribe(userAction => {
+      if (userAction == 0) {
+        //User has not performed any action on opened app dialog or closed the dialog;
+      } else if (userAction == 1) {
+        //User has approved delete operation 
+        this.spinnerService.setSpinnerVisibility(true);
+        this.assetmateService.deleteAsset(assetId).subscribe(res => {
+
+          this.spinnerService.setSpinnerVisibility(false);
+          this.showSnackBar(res.message);
+
+          this.getAllAssets(this.categoryID, this.page);
+
+        }, error => {
+          this.showSnackBar("Something went wrong..!!");
+        });
+      }
+    })
   }
 
   /*********************************************************** Edit Particular Asset  *******************************************************************/
